Update patch set 6
Patch Set 6:

> Can we please add a test that confirms that after installation of
 > this rpm Ansible is able to find this role?

We have two ways of doing it:

1. Testing it in the build environment in %check (we'd need to fix the modules path in ansible to match the one in the build environment)
2. Adding a new zuul job that installs the package from DLRN-rpmbuild-centos8 logs and checks it.

We usually don't create so specific jobs in zuul to gate changes in distgits but if maintainers think it may be worthy,  it may be a good idea.

In the meanwhile I'll send a review to add the test in %check.

Patch-set: 6
Label: Code-Review=0